Extract ContainerInfo in shared :utils:container-utils module
