+++ dd-java-agent/agent-profiling/profiling-uploader/profiling-uploader.gradle
@@ -16,6 +16,7 @@ excludedClassesCoverage += [
 
 dependencies {
   compile deps.slf4j
+  compile project(':utils:container-utils')
   compile project(':dd-trace-api')
 
   compile project(':dd-java-agent:agent-profiling:profiling-controller')
