+++ dd-trace-core/dd-trace-core.gradle
@@ -30,6 +30,7 @@ dependencies {
   compile project(':dd-trace-api')
   compile project(':internal-api')
   compile project(':utils:thread-utils')
+  compile project(':utils:container-utils')
 
   compile group: 'com.datadoghq', name: 'java-dogstatsd-client', version: "${versions.dogstatsd}"
