+++ settings.gradle
@@ -42,6 +42,7 @@ include ':dd-java-agent:agent-profiling:profiling-testing'
 include ':dd-java-agent:testing'
 include ':utils:test-utils'
 include ':utils:thread-utils'
+include ':utils:container-utils'
 
 // smoke tests
 include ':dd-smoke-tests:cli'
