+++ utils/container-utils/container-utils.gradle
@@ -1,0 +1,12 @@
+apply from: "$rootDir/gradle/java.gradle"
+
+// TODO: add more tests
+excludedClassesCoverage += [
+  'datadog.common.exec*'
+]
+
+dependencies {
+  compile deps.slf4j
+
+  testCompile project(':utils:test-utils')
+}
